Integrate coregister code into usfimr-s1 catalog folder
Add additional steps to main.sh

Reorganize python code a bit so the top level directory
isn't quite so cluttered.

Ran main.sh and uploaded catalog result to
s3://glofimr-sar-4326/glofimr-sar-catalog.zip
