--- catalogs/glofimr-s1/.gitignore
@@ -1,2 +1,0 @@
-data/*
-!data/.gitkeep
