+++ catalogs/usfimr-s1/.gitignore
@@ -1,1 +1,2 @@
-data/
+data/*
+!data/.gitkeep
