+++ catalogs/usfimr-s1/stac_utils/s3_io.py
